ai-deployer-engine/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
gunicorn>=21.2

slowapi>=0.1.9

pydantic>=2.6
pydantic-settings>=2.2

orjson>=3.9

python-dotenv>=1.0
structlog>=24.2.0
httpx>=0.27.0
